Implement multiple autosaves
Category: IO / UX

Rather than overwriting a single autosave over and over,
multiple autosaves can be made and cycled through.
E.g. as_2834_0000.blend, as_2834_0001.blend, as_2834_0002.blend
The user can control the number of autosaves in the preferences.
The default (1) keeps the old behavior, where a single autosave is made.
0 will keep making new autosaves indefinitely.

The counter for the next autosave will not change when the user changes
the max number of autosaves. I believe this behavior is better because
it will be less likely to overwrite, without needing to scan for files.