Sculpt: Enable Vector Displacement by default
Affects the default startup (the non-configured blender, or
when --factory-startup is used) and the "Reset Brush" operator.

This change does not modify existing brush settings in .blend
files as it is not possible to tell whether the option was intended
to be explicitly disabled or left at a default.

Caveat of the above is that adding new brushes does not enable the
option unless it was already enabled (since adding a brush actually
copies settings of the active brush).

Ref #106814
Ref #109992