WIP: makesrna don't regenerate files unless changed
makesrna is somewhat wasteful in causing rebuilds on files that
have not changed.

This makes 2 changes to makesrna.cc

1- the code in place to keep make out of trouble and regenerate files
regardless if they have changed or not has been made optional for the
MSVC and Ninja generators, as they do not appear to be needing this.

This brings the number of actions ninja takes down after a whitespace
change in rna_wm.cc from 423 down to 107

2- rna_prototypes_gen.h is always written, this is somewhat wasteful
and has been changed to be only written when changed.

This brings the number of actions ninja takes down after a whitespace
change in rna_wm.cc from 107 down to 11

in total, this saves several minutes of build time for simple changes
to rna source files.