Fix #110632: temp context doesn't always restore original area or region at exit.
Why exactly not explicitely restoring the context with `cxt_init`
values leaves relevant values to null is unclear to me? But this patch
seems to fix it, by considering that if a temp context value is set,
then it needs to be restored from the init value at the end, even if
both init and temp values are the same.