Cycles: Tweak Principled BSDF Subsurface parameters
Previously, the Principled BSDF used the Subsurface input to scale the radius.
When it was zero, it used a diffuse closure, otherwise a subsurface closure.

This sort of scaling input makes sense, but it should be specified in
distance units, rather than a 0..1 factor, so this commit changes the unit.

Additionally, it adds support for mixing diffuse and subsurface components.
This is part of e.g. the OpenPBR spec, and the logic behind it is to support
modeling e.g. dirt or paint on top of skin.

For typical materials, this will be either zero or one (like metallic or
transmission), but supporting fractional inputs makes sense for e.g.
smooth transitions at boundaries.