Fix #93103: modifier iterations can be set to a negative value
Sanitize "iterations" in a couple of modifiers (namely Smooth, Corrective
Smooth, Laplacian Smooth & Laplacian Deform) to always
have a hard limit of 0.

Putting in negative values does not make sense (and could have caused an
infinite loop since the negative value is cast to unsigned internally in
the case of the 'Corrective Smooth' modifier).