Fix #110833: Trackpad: Text Editor scrolling improvement
Improved Text Editor scrolling with the trackpad,
by accumulating deltas that are less than one line/character.

Makes the scroll speed match the speed in other Blender editors,
and other macOS applications as well.

The scrolling is very fast right now, deltas that gives less than
one line/character offset are discarded, it is almost impossible
to stop at a desired line.

This fix is originally authored Yevgeny Makarov (@jenkm) as
https://archive.blender.org/developer/D9552 , now porting
this as a PR.