Fix  #110848: Trackpad: Allow changing the speed while fly navigating
The code is already here, it was just lost during refactoring earlier.

This feature was added in rB4737e08cb9a5 and in rB3064da1200a9
keymap was lost, just because it wasn't added to the fly_modal_keymap
initially. This patch just adds FLY_MODAL_SPEED to modal map
so it can be selected in the keymap.

Also replace 'delta' with proper 'absolute delta'
so that the upward movement always increases the speed,
regardless of the Natural Scroll Direction preference.

This patch was originally authored by Yevgeny Makarov (@jenkm) as
https://archive.blender.org/developer/D9641 , now porting this as a PR.