Sim Nodes: Don't delete parent bake directory
The `OBJECT_OT_simulation_nodes_cache_delete` operator
would delete the parent bake directory. This can result in catastrophic
loss of data if the user set their bake directory to a folder that
contains other data.

This patch makes sure that only the "meta" and "bdata" folders get
deleted in the parent directory.