Subdiv: Use index instead of pointer for grid to face map
Halves the memory usage required for the `grid_faces` array. Using
indices should reduce required memory bandwidth, and allows better code
reuse with common utilities that are index-based. This map is the same
as the one created by `build_loop_to_face_map` for meshes, so the memory
should be shared in the future. The existing pointer array is often used with
subtraction to get the index anyway.