Cycles: Move Emission in Principled BSDF below Sheen and Coat
This way, sheen (e.g. a dust layer) and coat (e.g. a layer of paint) partially
obscure the emission and reduce its intensity.

In terms of implementation, this requires moving Emission, which previously
was expanded into a separate node, into the Principled node in SVM and OSL
so that we can compute the layering properly.