Cycles: Also move transparency into Principled node internally
For now, the transparency (which was also expanded out) could have remained
separate, but in the future when we e.g. implement the thin sheet mode, it'll
be helpful to handle transparency from within the node.

Note that there's optimization potential here - for fully transparent surfaces
and when evaluating transparent shadows, we should skip evaluating the other
components. Same goes for evaluating emission during light sampling.