Color management: Validate Look on View Transform changes
If the new view transform does not support the currently configured
look reset look to the default: the first look defined for the view.

Currently it is not a problem as all views support the same set of
looks. But with an upcoming addition of AgX it will no longer be
true.

Ref #110685