Realtime Compositor: Realize rotation for filter nodes
This patch realizes the rotation for the domain input of filter nodes.
That's because most filter nodes are not rotation invariant, so an input
with identity rotation is required for its correct operation in global
space.

This is not a real patch, but it was created to package builds for
testing and get some feedback.