Realtime Compositor: Realize rotation and scale for filter nodes
This patch realizes the rotation and scale for the domain input of
filter nodes. That's because most filter nodes are not rotation or scale
invariant, so an input with identity rotation and scale is required for
its correct operation in global space.

This is not a real patch, but it was created to package builds for
testing and get some feedback.

The patch does not handle the implications of huge and infinitesimal
textures resulted from the realization of scale.