Fix #109454: Snapping from Tools doesn't work if Depth option is not `Surface`
Tools and Drag & Drop snap options are edited through Tool Placement
settings.

This can cause confusion, even more so when the `Depth` option is not
`Surface`, in which case the Snap Toggle is simply ignored and appears
to be broken.

The solution then is to prioritize Snap Toggle over the `Depth` option,
so even if it's not `Surface`, the snap is still done when toggle is
enabled.