VSE: Ensure text offset x1 is never larger than x2
When you scroll all the way to the right side of the sequencer,
`x1` can become larger than `x2` due to the addition of
`scroller_vert_xoffs`.

When that happens, the assertion inside `UI_view2d_view_to_region_rcti_clip`
is triggered and causes blender to crash.

This change just clamps the value of x1 to the value of x2 to ensure the
assertion is never triggered.