PyGPU: Add Mipmap Support to gpu.types.GPUTexture
This PR adds mipmap support.

*Allocation*

When constructing a GPUTexture the max number of mipmap levels can
be set. The actual number of mipmap levels is adjusted based on the
size of the texture.

*Mipmap Generation*

*Setting filtermode*

NOTE: To actually move this patch forward we need to do some internal
restructurings and make a CPP Api of the GPU module.
Current API is to internal focussed.
So before continuing a design needs to be available for the new API.