Fix #108782: Fcurve Modifer shortcut does not work in channels region
Since the 5e1470d1b3eb / 0f51b5b599bb the `fmodifier_add` operator has
moved to the Channel menu in the Graph Editor and also the channels
context menu in the channels region. The shortcut was not working in the
channels region though since the operator was only in the keymap for the
main region (not the channels region).

Now move the keymap entry to the generic keymap of the Graph Editor so
it works in botch regions.