GPU: Swap Buffers
In the current implementation window manager directly calls GHOST
swap buffers. The GPU backend state could become invalid as it is
not aware of the buffer swap.

This PR will introduce GPU_context_swap_buffers so the backend can
do pre/post stuff to ensure the state is always valid.

* Reasoning *

In Vulkan the texture layout is not query-able, but the application is
responsible to always communicate the correct layout to the Vulkan API.
This isn't working for images from the swap chain.

In Vulkan the command buffer needs to be in a certain state as GHOST and
the GPU module share the same instance. Separating the command buffers
could be done, but would have synchronization issues as there can still
be command in flight or not, which isn't known by GHOST.

During start of blender contexts are initialized twice that leads to
incorrect behaviors as the state is not clear.
During start of blender buffers are swapped without submitting the
command buffer, leading to missing texture clears.

* Things to keep in mind *

- MTLBackend introduced begin/end frame. I think it is still required as
  swapbuffers are optional when using offscreen contexts