Revert "Fix #107185: Edit mode or existing attribute break rest position"
The commit that this reverts fixed the rest position attribute, which
wasn't added in edit mode, since the modifier evaluation is different.
The change was simple, but caused quite a few problems because it broke
the assumption in places that the evaluated result of the mesh modifier
stack in edit mode was also a BMesh rather than a mesh. In main this
has been resolved by a refactor, 91b27ab637334383b57b, but that change
is too complex to backport to 3.6. As discussed in the most recent nodes
module meeting, this reverts the original fix. That does mean that
the rest position attribute won't function properly in edit mode.

Fixes #110136, #110925, #109301, #109432

This reverts commit 82ca3d3604bded96f72650c23a7f95f29b956f30.