Subversion bump to avoid using DNA struct type check for versioning.
The `bNodeTreeInterface` DNA type has been added in main a while ago,
and new test files have been added in the meantime. This makes the
DNA check for versioning unreliable, it will skip versioning test files
which already have the DNA struct but need to be versioned.
Using a subversion bump is much more reliable.