Augmented socket callbacks with a simplified color function.
The new `draw_color_simple` callback does not depend on context or
a socket pointer or a node pointer, it only requires the socket
typeinfo (`bNodeSocketType`). In Python it's a registerable classmethod.
The existing `draw_color` callback also becomes optional in order to not
break existing scripts. One of the two functions should be implemented
by custom python nodes, preferably the `draw_color_simple` variant.