Fix registerable draw_color_simple callback, must not have SELF_TYPE.
This is very confusing RNA behavior: When a registerable method is
supposed to be a @classmethod it needs to have the `FUNC_NO_SELF` flag
set but NOT the `FUNC_USE_SELF_TYPE` flag (see
`bpy_class_validate_recursive` TODO comments). The overridden function
in the `NodeSocketStandard` subclass that has a built-in implementation,
however, needs the `FUNC_USE_SELF_TYPE` flag.